[FIX] make postsQuery Versatility
