+++ graphql/resolvers/Area/index.js
@@ -1,3 +1,5 @@
+const myAreaQuery = require("./myAreaQuery");
+
 const resolvers = {
   Query: {
     areas: async (_, __, { dataSources }) => {
@@ -8,37 +10,5 @@ const resolvers = {
     myArea: (_, args, { dataSources }) => myAreaQuery(args, dataSources),
   },
 };
-
-async function myAreaQuery({ lat, lng }, dataSources) {
-  const geoIntersectQuery = {
-    location: {
-      $near: {
-        $geometry: { type: "Point", coordinates: [lng, lat] },
-        $maxDistance: 0,
-      },
-    },
-  };
-
-  let area = await dataSources.areas.getAreaByQuery(geoIntersectQuery);
-
-  if (!area) {
-    const response = await dataSources.kakaoAPI.getRegions({ lat, lng });
-
-    if (!response || !response.documents || response.documents.length === 0) {
-      return null;
-    }
-
-    const region = response.documents[0];
-    const name = region.region_3depth_name || region.region_2depth_name || region.region_1depth_name;
-    const location = {
-      type: "Point",
-      coordinates: [region.x, region.y],
-    };
-
-    area = { name, location };
-  }
-
-  return area;
-}
 
 module.exports = resolvers;
